@@ -29,15 +29,15 @@
     "docker:push": "bash -c ./get_docker_image_name.sh | xargs -0 -I % docker push %"
   },
   "dependencies": {
-    "@angular/animations": "^5.2.10",
-    "@angular/common": "^5.2.5",
-    "@angular/compiler": "^5.2.5",
-    "@angular/core": "^5.2.5",
-    "@angular/forms": "^5.2.10",
-    "@angular/http": "^5.2.10",
-    "@angular/platform-browser": "^5.2.10",
-    "@angular/platform-browser-dynamic": "^5.2.10",
-    "@angular/router": "^5.2.10",
+    "@angular/animations": "6.0.7",
+    "@angular/common": "6.0.7",
+    "@angular/compiler": "6.0.7",
+    "@angular/core": "6.0.7",
+    "@angular/forms": "6.0.7",
+    "@angular/http": "6.0.7",
+    "@angular/platform-browser": "6.0.7",
+    "@angular/platform-browser-dynamic": "6.0.7",
+    "@angular/router": "6.0.7",
     "@fortawesome/fontawesome-free-webfonts": "^1.0.9",
     "@ng-bootstrap/ng-bootstrap": "^1.1.2",
     "@ngx-translate/core": "^9.1.1",
@@ -45,48 +45,50 @@
     "babel-cli": "^6.26.0",
     "babel-preset-es2015": "^6.24.1",
     "bootstrap": "4.0.0",
-    "core-js": "^2.5.2",
+    "core-js": "^2.5.7",
     "jquery": "^3.3.1",
     "lodash": "^4.17.4",
     "ng-scrollreveal": "^2.2.0",
     "popper.js": "^1.14.3",
-    "rxjs": "^5.5.5",
+    "rxjs": "^6.2.1",
+    "rxjs-compat": "^6.2.1",
     "scrollreveal": "^3.4.0",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.4",
-    "@angular/compiler-cli": "^5.2.10",
-    "@angular/language-service": "^5.2.10",
-    "@angular/platform-server": "^6.0.3",
+    "@angular-devkit/build-angular": "~0.6.8",
+    "@angular/cli": "6.0.8",
+    "@angular/compiler-cli": "6.0.7",
+    "@angular/language-service": "6.0.7",
+    "@angular/platform-server": "6.0.7",
     "@biesbjerg/ngx-translate-extract": "^2.3.1",
     "@ngx-rocket/scripts": "^2.2.0",
-    "@types/jasmine": "^2.8.7",
+    "@types/jasmine": "^2.8.8",
     "@types/jasminewd2": "^2.0.2",
-    "@types/lodash": "^4.14.108",
-    "@types/node": "^8.10.15",
+    "@types/lodash": "^4.14.110",
+    "@types/node": "^8.10.20",
     "@types/scrollreveal": "0.0.3",
-    "codelyzer": "^4.3.0",
+    "codelyzer": "^4.4.2",
     "hads": "^1.3.4",
     "htmlhint": "^0.9.13",
     "https-proxy-agent": "^2.2.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "^4.1.0",
-    "karma": "^2.0.2",
+    "karma": "^2.0.4",
     "karma-chrome-launcher": "^2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
+    "karma-coverage-istanbul-reporter": "^1.4.3",
     "karma-jasmine": "^1.1.2",
     "karma-junit-reporter": "^1.2.0",
     "protractor": "^5.3.2",
-    "puppeteer": "^1.4.0",
+    "puppeteer": "^1.5.0",
     "replace-in-file": "^3.4.0",
-    "stylelint": "^9.2.0",
+    "stylelint": "^9.3.0",
     "stylelint-config-recommended-scss": "~3.1.0",
     "stylelint-config-standard": "~18.1.0",
     "stylelint-scss": "~2.4.0",
     "ts-node": "^5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~2.6.2"
+    "typescript": "2.7.2"
   }
 }
